Sound changes and particle battery
On full build, tank shots weren't making noise I have changed the code for it to only have one main tank shot and removed the dynamics until the rest of sound is working. I have also updated the particle on the battery for it to more contained and look like the battery pulse itself.